+++ app/main/views/add_service.py
@@ -6,7 +6,7 @@ from app import service_api_client
 from app.formatters import email_safe
 from app.main import main
 from app.main.forms import CreateNhsServiceForm, CreateServiceForm
-from app.utils import user_is_gov_user, user_is_logged_in
+from app.utils.user import user_is_gov_user, user_is_logged_in
 
 
 def _create_service(service_name, organisation_type, email_from, form):
