+++ app/main/views/broadcast.py
@@ -19,7 +19,8 @@ from app.main.forms import (
     SearchByNameForm,
 )
 from app.models.broadcast_message import BroadcastMessage, BroadcastMessages
-from app.utils import service_has_permission, user_has_permissions
+from app.utils import service_has_permission
+from app.utils.user import user_has_permissions
 
 
 def _get_back_link_from_view_broadcast_endpoint():
