+++ app/main/views/dashboard.py
@@ -35,8 +35,8 @@ from app.utils import (
     generate_previous_dict,
     get_current_financial_year,
     service_has_permission,
-    user_has_permissions,
 )
+from app.utils.user import user_has_permissions
 
 
 @main.route("/services/<uuid:service_id>/dashboard")
