+++ app/main/views/history.py
@@ -6,7 +6,7 @@ from flask import render_template, request
 from app import current_service, format_date_numeric
 from app.main import main
 from app.models.event import APIKeyEvent, APIKeyEvents, ServiceEvents
-from app.utils import user_has_permissions
+from app.utils.user import user_has_permissions
 
 
 @main.route("/services/<uuid:service_id>/history")
