+++ app/main/views/returned_letters.py
@@ -5,7 +5,7 @@ from flask import render_template
 from app import current_service, service_api_client
 from app.main import main
 from app.models.spreadsheet import Spreadsheet
-from app.utils import user_has_permissions
+from app.utils.user import user_has_permissions
 
 
 @main.route("/services/<uuid:service_id>/returned-letters")
