+++ app/main/views/webauthn_credentials.py
@@ -10,11 +10,8 @@ from app.main.views.two_factor import log_in_user
 from app.models.user import User
 from app.models.webauthn_credential import RegistrationError, WebAuthnCredential
 from app.notify_client.user_api_client import user_api_client
-from app.utils import (
-    is_less_than_days_ago,
-    redirect_to_sign_in,
-    user_is_platform_admin,
-)
+from app.utils import is_less_than_days_ago, redirect_to_sign_in
+from app.utils.user import user_is_platform_admin
 
 
 @main.route('/webauthn/register')
