+++ tests/app/main/views/test_add_service.py
@@ -3,7 +3,7 @@ from flask import session, url_for
 from freezegun import freeze_time
 from notifications_python_client.errors import HTTPError
 
-from app.utils import is_gov_user
+from app.utils.user import is_gov_user
 from tests import organisation_json
 from tests.conftest import normalize_spaces
